add: new endpoint from create sale

+++ server/src/products/models/product_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, Numeric, String
 from server.src.data.database import Base
 
 
@@ -7,6 +7,6 @@ class Product(Base):
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String, nullable=False)
-    price = Column(Float, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
     stock = Column(Integer, default=0)
     barcode = Column(String, nullable=False)
